extract more *.img files

+++ oe/branches/branch-0-0-1/oe/packages/samsung/firmware-package.bbclass
@@ -6,23 +6,36 @@ do_patch () {
 	cd "${S}"
 	ls -l ./
 	echo "### Big fat warning! sh4 (& unencoded appdata + exe img) arch is to implement!!"
+	SEC="no" ; ENC="no"
 	for i in exe.img appdata.img ; do
 		# cip decrypt if .sec file present
 		if [ -e ${MACHINE}/image/$i.sec ] ; then
+		# notice that file was sec
+		SEC="yes"
 		# mprotect's decrypt_update tool
 		#	decrypt_update ${MACHINE}/image/$i.sec ${MACHINE}/image/$i
 		# or decrypt with openssl
 			ENDE=$(stat ${MACHINE}/image/$i.sec | grep Size: | cut -d ":" -f2 | cut -f1)
 			SHORT=$((${ENDE} - 260))
 			dd if=${MACHINE}/image/$i.sec of=${MACHINE}/image/$i.sec.cut bs=${SHORT} count=1	
-			openssl aes-128-cbc -d -in ${MACHINE}/image/$i.sec.cut -out ${MACHINE}/image/$i.enc -pass pass:7cb49e2a2fb8c4dd1df3b41e20cbb6179c8bd240
+			openssl aes-128-cbc -d -in ${MACHINE}/image/$i.sec.cut -out ${MACHINE}/image/$i.enc \
+				-pass pass:7cb49e2a2fb8c4dd1df3b41e20cbb6179c8bd240
 		fi
+		# x-or .enc file if present
+		if [ -e ${MACHINE}/image/$i.enc ] ; then
+		# notice that file was x-ored
+		ENC="yes"
 		crypt-xor -f "${MACHINE}/image/$i.enc" -K "${MACHINE}" -force -q -outfile "${MACHINE}/image/$i"
 		rm -f "${MACHINE}/image/$i.enc"
+		fi
 	# extract files for later inspection
-	unsquashfs ${MACHINE}/image/$i || $(mkdir tt && mount -o loop,noatime ${MACHINE}/image/$i tt ; cp -a tt/* ./ ; umount tt ; rm -rf tt)
+	mkdir -p ${MACHINE}-orig
+	unsquashfs -dest ${MACHINE}-orig/$i ${MACHINE}/image/$i || $(mkdir tt && mount -o loop,noatime ${MACHINE}/image/$i tt ; cp -af tt/* ${MACHINE}-orig/ ; umount tt ; rm -rf tt)
 	done
-	crypt-xor -f "${MACHINE}/run.sh.enc" -K "${MACHINE}" -force -q -outfile run.sh || true
+	# we can ignore this file, its never used
+	# crypt-xor -f "${MACHINE}/run.sh.enc" -K "${MACHINE}" -force -q -outfile ${MACHINE}-orig/run.sh || true
+	if [ -n "${P_LINE}" ] ; then
+	oedebug "P_LINE Present, so patch"
 	# method 1.
 	# mkdir tt && mount -o loop,noatime exe.img tt 
 	# change whata u want
@@ -30,17 +43,33 @@ do_patch () {
 	# method 2.
 	echo "${P_LINE}" | dd conv=notrunc of="${MACHINE}/image/exe.img" seek=${P_OFFSET} bs=1
 	# end of methods
+	fi
+	# sh4 has this too? we regenerate the file if it was present
+	if [ -e ${MACHINE}/image/validinfo.txt ] ; then
 	cksfv -b ${MACHINE}/image/*.img > ${MACHINE}/image/validinfo.txt
+	fi
 	for i in exe.img appdata.img ; do
+		mkdir -p ${MACHINE}-patched
+		unsquashfs -dest ${MACHINE}-patched/$i ${MACHINE}/image/$i  || (mkdir -p tt && mount -o loop,noatime ${MACHINE}/image/$i tt ; cp -a tt/* ${MACHINE}-patched/ ; umount tt ; rm -rf tt)
+		if [ $ENC = "yes" ] ; then
+		oedebug "Firmware was x-ored, regenerate it"
                 crypt-xor -f "${MACHINE}/image/$i" -K "${MACHINE}" -force -q -outfile "${MACHINE}/image/$i.enc"
+		fi
 		if [ -e ${MACHINE}/image/$i.sec ] ; then
-			openssl aes-128-cbc -e -in ${MACHINE}/image/$i.enc -out ${MACHINE}/image/$i.sec.nosig -pass pass:7cb49e2a2fb8c4dd1df3b41e20cbb6179c8bd240 -debug
+		oedebug "Firmware was for ci+, encrypt it again"
+		oenote "Don't try to flash it, signature generation method unknown. Image isn't walid!!!!"
+			openssl aes-128-cbc -e -in ${MACHINE}/image/$i.enc -out ${MACHINE}/image/$i.sec.nosig \
+				-pass pass:7cb49e2a2fb8c4dd1df3b41e20cbb6179c8bd240
 		fi
 #		rm -f "${MACHINE}/image/$i"
-		mv "${MACHINE}/image/$i" ./
-		mkdir -p ./Patched || true
-		unsquashfs -dest ./Patched ./$i || $(mkdir -p ./tt && mount -o loop,noatime ./$i ./tt ; cp -a tt/* ./Patched/ ; umount tt ; rm -rf tt)
+#		mv "${MACHINE}/image/$i" ./
         done
+	# extract some extra image files
+	for i in cmm.img chip.img boot.img rootfs.img ; do
+	if [ -e ${MACHINE}/image/$i ] ; then
+		unsquashfs -f -dest ${MACHINE}-orig/ ${MACHINE}/image/$i || $(mkdir tt && mount -o loop,noatime ${MACHINE}/image/$i tt ; cp -af tt/* ${MACHINE}-orig/ ; umount tt ; rm -rf tt)
+	fi
+	done
 	
 }
